Update README.md
This was a final project for a class in statistical modeling and analysis. The class was tasked with finding a publically available dataset and making models to solve a fairly simple problem with business applications. My group and I chose this dataset from Kaggle: https://www.kaggle.com/jessemostipak/hotel-booking-demand. For us, this seemed to be a suitable dataset, as the business application was developing a model to predict hotel cancellations. We began with an initial logistic regression, then moved on to Ridge, Lasso, and ElasticNet models. We concluded with two random forests, one that included a specific selection of variables, and another that used all the variables we thought relevant and applicable. You will find our RMarkdown document, the CSV file, and our final presentation explaining it. Feel free to work on it or offer feedback!"